requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.13
ollama>=0.1.0

# OCR and image processing
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3
opencv-python>=4.8.0

# Document processing
pypdf>=3.17.0
python-docx>=1.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
